+++ src/parse.rs
@@ -99,18 +99,18 @@ impl Parse {
     ///
     /// If the next entry cannot be represented as an integer, then an error is
     /// returned.
-    pub(crate) fn next_int(&mut self) -> Result<u64, ParseError> {
+    pub(crate) fn next_int(&mut self) -> Result<i64, ParseError> {
         use atoi::atoi;
 
         const MSG: &str = "protocol error; invalid number";
 
         match self.next()? {
             // An integer frame type is already stored as an integer.
-            Frame::Integer(v) => Ok(v as u64),
+            Frame::Integer(v) => Ok(v),
             // Simple and bulk frames must be parsed as integers. If the parsing
             // fails, an error is returned.
-            Frame::Simple(data) => atoi::<u64>(data.as_bytes()).ok_or_else(|| MSG.into()),
-            Frame::Bulk(data) => atoi::<u64>(&data).ok_or_else(|| MSG.into()),
+            Frame::Simple(data) => atoi::<i64>(data.as_bytes()).ok_or_else(|| MSG.into()),
+            Frame::Bulk(data) => atoi::<i64>(&data).ok_or_else(|| MSG.into()),
             frame => Err(format!("protocol error; expected int frame but got {:?}", frame).into()),
         }
     }
